Move Organization models to sub namespaces
