+++ src/Api/Models/Request/Organizations/OrganizationUserRequestModels.cs
@@ -6,6 +6,7 @@ using System.Text.Json;
 using Bit.Core.Entities;
 using Bit.Core.Enums;
 using Bit.Core.Models.Data;
+using Bit.Core.Models.Data.Organizations.OrganizationUsers;
 using Bit.Core.Utilities;
 
 namespace Bit.Api.Models.Request.Organizations
