+++ src/Api/Models/Response/Organizations/OrganizationAutoEnrollStatusResponseModel.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using Bit.Core.Models.Api;
 
-namespace Bit.Api.Models.Response
+namespace Bit.Api.Models.Response.Organizations
 {
     public class OrganizationAutoEnrollStatusResponseModel : ResponseModel
     {
