+++ src/Api/Models/Response/Organizations/OrganizationKeysResponseModel.cs
@@ -2,7 +2,7 @@
 using Bit.Core.Entities;
 using Bit.Core.Models.Api;
 
-namespace Bit.Api.Models.Response
+namespace Bit.Api.Models.Response.Organizations
 {
     public class OrganizationKeysResponseModel : ResponseModel
     {
