+++ src/Api/Models/Response/ProfileOrganizationResponseModel.cs
@@ -2,6 +2,7 @@
 using Bit.Core.Enums;
 using Bit.Core.Models.Api;
 using Bit.Core.Models.Data;
+using Bit.Core.Models.Data.Organizations.OrganizationUsers;
 using Bit.Core.Utilities;
 
 namespace Bit.Api.Models.Response
