+++ src/Api/Models/Response/SyncResponseModel.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using Bit.Core.Entities;
 using Bit.Core.Models.Api;
 using Bit.Core.Models.Data;
+using Bit.Core.Models.Data.Organizations.OrganizationUsers;
 using Bit.Core.Settings;
 using Core.Models.Data;
