+++ src/Core/Models/Data/Organizations/OrganizationUsers/OrganizationUserInviteData.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
 using Bit.Core.Enums;
 
-namespace Bit.Core.Models.Data
+namespace Bit.Core.Models.Data.Organizations.OrganizationUsers
 {
     public class OrganizationUserInviteData
     {
