+++ src/Core/Models/Data/Organizations/OrganizationUsers/OrganizationUserPublicKey.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace Bit.Core.Models.Data
+namespace Bit.Core.Models.Data.Organizations.OrganizationUsers
 {
     public class OrganizationUserPublicKey
     {
