+++ src/Core/Models/Data/Organizations/OrganizationUsers/OrganizationUserWithCollections.cs
@@ -1,7 +1,7 @@
 ﻿using System.Data;
 using Bit.Core.Entities;
 
-namespace Bit.Core.Models.Data
+namespace Bit.Core.Models.Data.Organizations.OrganizationUsers
 {
     public class OrganizationUserWithCollections : OrganizationUser
     {
