+++ src/Core/Models/StaticStore/SponsoredPlan.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 using Bit.Core.Enums;
-using Bit.Core.Models.Data;
+using Bit.Core.Models.Data.Organizations.OrganizationUsers;
 
 namespace Bit.Core.Models.StaticStore
 {
