+++ src/Core/Services/Implementations/InMemoryApplicationCacheService.cs
@@ -5,6 +5,7 @@ using System.Threading.Tasks;
 using Bit.Core.Entities;
 using Bit.Core.Entities.Provider;
 using Bit.Core.Models.Data;
+using Bit.Core.Models.Data.Organizations;
 using Bit.Core.Repositories;
 
 namespace Bit.Core.Services
